Add files via upload
I have put together Yan Yan's code and my code as well as some of Justin's into a sort of a working Menu to access things. The files needed to use in xcode have been uploaded. This code works and can be compiled. Yan Yan's code is clean and is very nice, Justin's as well. This is what I have done so far.

NOTE:
To view the .txt information after compiling code, set up the path in xcode: Product >> Scheme >> Edit Scheme >> Then check the "Working Directory" box and put in a place where you can locate it easily in Finder.