notes on current code
Hey guys, I was really blown away with what YanYan had done and Justin's code and wanted to find a way to combine a template for what we have so far. I see clearly how much hard work it all was. You guys are awesome. I tried to make a master of all of our code put together and the files i uploaded today are the result. They are still in a work in progress, and I will be updating them as I try to develop more this weekend and next week. If anyone else has developed code please feel free to upload and I will try to combine them as well. Also, if any of you don't want the formatting this way I am of course open to redesign:)